Add files via upload
Project Requirements: Sign-Up Application

we're going to be writing a user registration application. This will allow users to register with the system, providing both their name and age, and we'll store this information in our own custom type. We'll also provide the ability for a user to be looked up by an ID, retrieving their information.

Once you complete the activity, you should obtain an output similar to the following uploaded image .